+++ src/main/java/kr/or/cola/backend/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package kr.or.cola.backend.config;
 
-import kr.or.cola.backend.config.auth.OAuth2AuthenticationSuccessHandler;
+import kr.or.cola.backend.oauth.OAuth2AuthenticationSuccessHandler;
 import kr.or.cola.backend.user.UserService;
 import kr.or.cola.backend.user.domain.Role;
 import lombok.RequiredArgsConstructor;
