+++ src/main/java/kr/or/cola/backend/config/WebConfig.java
@@ -1,6 +1,6 @@
 package kr.or.cola.backend.config;
 
-import kr.or.cola.backend.config.auth.LoginUserArgumentResolver;
+import kr.or.cola.backend.oauth.LoginUserArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
