+++ src/main/java/kr/or/cola/backend/oauth/LoginUserArgumentResolver.java
@@ -1,6 +1,6 @@
-package kr.or.cola.backend.config.auth;
+package kr.or.cola.backend.oauth;
 
-import kr.or.cola.backend.config.auth.dto.SessionUser;
+import kr.or.cola.backend.oauth.dto.SessionUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
